Fix bug where 'erase-buffer' is called too late
The call to 'erase-buffer' was done after the process had started,
which had the effect of deleting the entire results.  I tried to add
the 'erase-buffer' to a :filter of 'make-process', but then there was
a similar problem where the function was called after the process had
started but before its conclusion, thus deleting most results though
not all of them.

Calling 'erase-buffer' before 'make-process' ensures that we are
always getting the same results, while being able to re-use an
existing buffer (such as by reverting it by typing 'g').